+++ ydb/library/workload/query/query.h
@@ -12,7 +12,7 @@ public:
     THolder<IWorkloadQueryGenerator> CreateGenerator() const override;
     TWorkloadDataInitializer::TList CreateDataInitializers() const override;
     TString GetWorkloadName() const override;
-    TString GetDescription() const override;
+    TString GetDescription(ECommandType commandType, int workloadType) const override;
 
     YDB_READONLY_DEF(TFsPath, SuitePath);
     YDB_READONLY_DEF(TVector<TString>, CustomQueries);